+++ apps/web/src/features/auth/pages/landing.tsx
@@ -1,16 +1,16 @@
-import { useResolve, useNavigate } from '@viness/react'
+import { useResolve, useNavigate } from '@viness/react';
 
 export function LandingPage() {
-    const navigate = useNavigate()
+    const navigate = useNavigate();
 
     const toPage = () => {
-        navigate('/dashboard')
-    }
+        navigate('/dashboard');
+    };
 
     return (
         <div>
             LandingPage
             <button onClick={toPage}>test</button>
         </div>
-    )
+    );
 }