+++ apps/web/src/features/dashboard/store/counter-store.hook.ts
@@ -1,4 +1,0 @@
-import { useResolve } from '@viness/react';
-import { CounterStore } from './counter-store';
-
-export const useCounterStore = () => useResolve<CounterStore>(CounterStore);