+++ packages/viness-react/src/index.tsx
@@ -1,3 +1,5 @@
+import '@abraham/reflection';
+
 export * from './app';
 
 export * from './hooks';