+++ packages/viness-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @viness/core
 
+## 1.5.1
+
+### Patch Changes
+
+- feat: change store to state & simplify the use of ui state
+- Updated dependencies
+  - @viness/core@1.5.1
+
 ## 1.5.0
 
 ### Minor Changes